Agregue las funciones para los controladores a una carpeta separada llamada helpers
